refactor(upcoming): extract upcoming episode item
